Rename react component files to CamelCase
